// src/components/sounds/type.ts
export type SoundName =
  | 'bgm'
  | 'yiqi'
  | 'chi'
  | 'hao'
  | 'click'
  | 'button'
  | 'miss'
  | 'levelup'
  | 'success'
  | 'correct'
  | 'incorrect';

export type SoundVOName =
  | 'sound1'
  | 'sound2'
  | 'sound3'
  | 'sound4'
  | 'sound5'
  | 'sound6'
  | 'sound7'
  | 'sound8'
  | 'sound9'
  | 'sound10'
  | 'sound11'
  | 'sound12'
  | 'sound13'
  | 'sound14'
  | 'sound15'
  | 'sound16'
  | 'sound17'
  | 'sound18'
  | 'sound19'
  | 'sound20'
  | 'sound21'
  | 'sound22'
  | 'sound23';

export type SoundHeyLongName = 'heyLong1' | 'heyLong2';